added the 2018 Brand Yourself presentation
added the new blog post. changed the redirect (/brand-yourself) to the new page. Added links from the old page to the new page and vice versa. Renamed the old page from branding_yourself to brand_yourself_1. Added this to CV. Added the flyer. Added it to "recent presentations" list and removed NWAV46.